// packages/datadog-api-client-v1/models/WidgetLayout.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* The layout for a widget on a `free` or **new dashboard layout** dashboard.
*/
export declare class WidgetLayout {
    /**
    * The height of the widget. Should be a non-negative integer.
    */
    'height': number;
    /**
    * Whether the widget should be the first one on the second column in high density or not. **Note**: Only for the **new dashboard layout** and only one widget in the dashboard should have this property set to `true`.
    */
    'isColumnBreak'?: boolean;
    /**
    * The width of the widget. Should be a non-negative integer.
    */
    'width': number;
    /**
    * The position of the widget on the x (horizontal) axis. Should be a non-negative integer.
    */
    'x': number;
    /**
    * The position of the widget on the y (vertical) axis. Should be a non-negative integer.
    */
    'y': number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): WidgetLayout;
    static serialize(data: WidgetLayout): {
        [key: string]: any;
    };
    constructor();
}